Fixed a few compilation warnings in the contrib folder and added a check against each particle index in the COM force.

+++ contrib/rovigatti/GraftedInteraction.cpp
@@ -77,7 +77,7 @@ void GraftedInteraction<number>::allocate_particles(BaseParticle<number> **parti
 	}
 	fclose(af);
 
-	if(anchors.size() != _N_arms) throw oxDNAException("The anchor file '%s' contains only %d positions while, according to the topology, there should be %d", anchors.size(), _N_arms);
+	if((int)anchors.size() != _N_arms) throw oxDNAException("The anchor file '%s' contains only %d positions while, according to the topology, there should be %d", anchors.size(), _N_arms);
 
 	particles[0] = new Colloid<number>(anchors);
